Displays tiles
